+++ CHANGELOG.md
@@ -4,6 +4,8 @@
    - Unify the build process by adding git-submodule metrics-base in
      src/build and removing all centralized parts of the build
      process.
+   - Accept the same characters in qualified search terms as in their
+     parameter equivalents.
 
 
 # Changes in version 3.1-1.0.0 - 2016-11-23